Don't provide 'uint64_t' as POPT_ARG_LONG.
Sizes might not match on some platforms.

Resolves: https://github.com/SSSD/sssd/issues/6790

+++ src/providers/proxy/proxy_child.c
@@ -538,7 +538,7 @@ int main(int argc, const char *argv[])
     debug_log_file = talloc_asprintf(NULL, "proxy_child_%s", domain);
     if (!debug_log_file) return 2;
 
-    sss_chain_id_set(chain_id);
+    sss_chain_id_set((uint64_t)chain_id);
 
     DEBUG_INIT(debug_level, opt_logger);
